pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bookmark-research-engine"
version = "0.1.0"
description = "Turn X bookmarks into ranked, analyzed research briefs"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.scripts]
bre = "bre.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bre"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
